chore(broadcast): move IntentReceiver to broadcast, convert it to Kotlin
Java cannot utilizes Kotlin's Coroutine well actually, so we need to convert it.
